+++ crates/router/src/connector/bankofamerica/transformers.rs
@@ -2600,7 +2600,9 @@ impl From<BankofamericaRefundStatus> for enums::RefundStatus {
             BankofamericaRefundStatus::Succeeded | BankofamericaRefundStatus::Transmitted => {
                 Self::Success
             }
-            BankofamericaRefundStatus::Failed | BankofamericaRefundStatus::Voided => Self::Failure,
+            BankofamericaRefundStatus::Cancelled
+            | BankofamericaRefundStatus::Failed
+            | BankofamericaRefundStatus::Voided => Self::Failure,
             BankofamericaRefundStatus::Pending => Self::Pending,
         }
     }
@@ -2638,6 +2640,7 @@ pub enum BankofamericaRefundStatus {
     Failed,
     Pending,
     Voided,
+    Cancelled,
 }
 
 #[derive(Debug, Deserialize, Serialize)]